Move test into own Test Case and Kernel
