+++ src/Sulu/Bundle/MediaBundle/Tests/Application/config/config_secured.yml
@@ -1,0 +1,24 @@
+security:
+    enable_authenticator_manager: true
+
+    access_decision_manager:
+        strategy: affirmative
+
+    encoders:
+        Sulu\Bundle\SecurityBundle\Entity\User: plaintext
+
+    providers:
+        testprovider:
+            id: test_user_provider
+
+    firewalls:
+        test:
+            lazy: true
+            http_basic:
+
+sulu_test:
+    enable_test_user_provider: true
+
+sulu_security:
+    checker:
+        enabled: true
