+++ src/Sulu/Bundle/MediaBundle/Tests/Functional/Controller/MediaStreamControllerTest.php
@@ -110,72 +110,6 @@ class MediaStreamControllerTest extends WebsiteTestCase
         );
     }
 
-    public function testDownloadWithCollectionPermissions(): void
-    {
-        $filePath = $this->createMediaFile('test.jpg');
-        $media = $this->createMedia($filePath, 'file-without-extension');
-
-        $anonymousRole = $this->getContainer()->get('sulu.repository.role')->createNew();
-        $this->getContainer()->get('doctrine.orm.entity_manager')->persist($anonymousRole);
-        $anonymousRole->setName('Anonymous');
-        $anonymousRole->setAnonymous(true);
-        $anonymousRole->setSystem('sulu_io');
-
-        $allowedContact = new Contact();
-        $this->getContainer()->get('doctrine.orm.entity_manager')->persist($allowedContact);
-        $allowedContact->setFirstName('Allowed');
-        $allowedContact->setLastName('Contact');
-
-        $allowedUser = new User();
-        $this->getEntityManager()->persist($allowedUser);
-        $allowedUser->setUsername('allowed-user');
-        $allowedUser->setContact($allowedContact);
-        $allowedUser->setSalt('');
-
-        $encoder = self::$container->get('security.encoder_factory')->getEncoder($allowedUser);
-        $allowedUser->setPassword($encoder->encodePassword('allowed-user', $allowedUser->getSalt()));
-        $allowedUser->setLocale('en');
-
-        $allowedRole = $this->getContainer()->get('sulu.repository.role')->createNew();
-        $this->getContainer()->get('doctrine.orm.entity_manager')->persist($allowedRole);
-        $allowedRole->setName('Allowed Role');
-        $allowedRole->setAnonymous(false);
-        $allowedRole->setSystem('sulu_io');
-
-        $allowedUserRole = new UserRole();
-        $this->getContainer()->get('doctrine.orm.entity_manager')->persist($allowedUserRole);
-        $allowedUserRole->setRole($allowedRole);
-        $allowedUserRole->setUser($allowedUser);
-        $allowedUserRole->setLocale(\json_encode(['de', 'en']) ?: '');
-        $allowedUser->addUserRole($allowedUserRole);
-
-        $this->getContainer()->get('doctrine.orm.entity_manager')->flush();
-
-        $this->getContainer()->get('sulu_security.access_control_manager')->setPermissions(
-            Collection::class,
-            (string) $media->getCollection(),
-            [
-                $anonymousRole->getId() => [
-                    'view' => false,
-                ],
-                $allowedRole->getId() => [
-                    'view' => true,
-                ],
-            ]
-        );
-
-        $this->client->jsonRequest('GET', 'http://sulu.lo' . $media->getUrl());
-        $unauthenticatedResponse = $this->client->getResponse();
-        $this->assertHttpStatusCode(401, $unauthenticatedResponse);
-
-        $this->client->jsonRequest('GET', 'http://sulu.lo' . $media->getUrl(), [], [
-            'PHP_AUTH_USER' => 'allowed-user',
-            'PHP_AUTH_PW' => 'allowed-user',
-        ]);
-        $authenticatedResponse = $this->client->getResponse();
-        $this->assertHttpStatusCode(200, $authenticatedResponse);
-    }
-
     public function testGetImageActionForNonExistingMedia()
     {
         $this->client->jsonRequest('GET', '/uploads/media/sulu-400x400/01/test.jpg?v=1');
